requirements.txt
fastapi
fastapi_offline
neo4j
requests
python-dotenv
python-docx
PyPDF2
uvicorn
blingfire
urllib3
pydantic
elasticsearch
scipy
# Efficient NLP libraries
spacy>=3.7.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
# Vector search optimization
faiss-cpu>=1.7.4
# Enhanced async support
httpx>=0.25.0
# Caching and performance
redis>=4.5.0
# Testing and logging
pytest>=7.4.0
pytest-asyncio>=0.21.0
